third_party/blink/web_tests/external/wpt/css/css-text-decor/parsing/text-decoration-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property text-decoration value 'none'
  assert_equals: expected "rgb(0, 0, 255)" but got "none solid rgb(0, 0, 255)"
[FAIL] Property text-decoration value 'line-through'
  assert_equals: expected "line-through rgb(0, 0, 255)" but got "line-through solid rgb(0, 0, 255)"
[FAIL] Property text-decoration value 'solid'
  assert_equals: expected "rgb(0, 0, 255)" but got "none solid rgb(0, 0, 255)"
[FAIL] Property text-decoration value 'currentcolor'
  assert_equals: expected "rgb(0, 0, 255)" but got "none solid rgb(0, 0, 255)"
[FAIL] Property text-decoration value 'underline overline line-through red'
  assert_equals: expected "underline overline line-through rgb(255, 0, 0)" but got "underline overline line-through solid rgb(255, 0, 0)"
[FAIL] Property text-decoration value 'rgba(10, 20, 30, 0.4) dotted'
  assert_equals: expected "dotted rgba(10, 20, 30, 0.4)" but got "none dotted rgba(10, 20, 30, 0.4)"
[FAIL] Property text-decoration value 'auto'
  assert_equals: expected "rgb(0, 0, 255)" but got "none solid rgb(0, 0, 255)"
[FAIL] Property text-decoration value 'from-font'
  assert_equals: expected "from-font rgb(0, 0, 255)" but got "none from-font solid rgb(0, 0, 255)"
[FAIL] Property text-decoration value '10px'
  assert_equals: expected "10px rgb(0, 0, 255)" but got "none 10px solid rgb(0, 0, 255)"
[FAIL] Property text-decoration value 'underline red from-font'
  assert_equals: expected "underline from-font rgb(255, 0, 0)" but got "underline from-font solid rgb(255, 0, 0)"
Harness: the test ran to completion.
